+++ internal/riverinternaltest/riverinternaltest.go
@@ -26,6 +26,16 @@ import (
 	"github.com/riverqueue/river/internal/util/valutil"
 )
 
+// SchedulerShortInterval is an artificially short interval for the scheduler
+// that's used in the tests of various components to make sure that errored jobs
+// always end up in a `retryable` state rather than `available`. Normally, the
+// job executor sets `available` if the retry delay is smaller than the
+// scheduler's interval. To simplify things so errors are always `retryable`,
+// this time is picked to be smaller than the any retry delay that the default
+// retry policy will ever produce. It's shared so we can document/explain it all
+// in one place.
+const SchedulerShortInterval = 500 * time.Millisecond
+
 var (
 	dbManager *testdb.Manager //nolint:gochecknoglobals